+++ packages/orq-rc/RELEASES.md
@@ -3829,3 +3829,13 @@ Based on:
 - [typescript v3.13.0-rc.29] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.29 - packages/orq-rc
+
+## 2025-09-24 04:56:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.624.0 (2.714.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.30] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.30 - packages/orq-rc